requirements.txt
numpy>=1.26
opencv-python-headless>=4.8
pyscenedetect>=0.6; python_version < "3.13"
moderngl>=5.8
moviepy>=1.0.3
PyOpenGL>=3.1.6
tqdm>=4.66
pydantic>=2.5
PyYAML>=6.0
